+++ Libplanet.Tests/Common/Action/BattleResult.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace Libplanet.Tests.Common.Action
 {
     [Serializable]
-    public class BattleResult
+    public class BattleResult : IEquatable<BattleResult>
     {
         public BattleResult()
         {
@@ -15,5 +15,27 @@ namespace Libplanet.Tests.Common.Action
         public HashSet<string> UsedWeapons { get; internal set; }
 
         public HashSet<string> Targets { get; internal set; }
+
+        public override bool Equals(object other)
+        {
+            return other != null && other is IEquatable<BattleResult> o &&
+                   o.Equals(this);
+        }
+
+        public bool Equals(BattleResult other)
+        {
+            return other != null &&
+                UsedWeapons.SetEquals(other.UsedWeapons) &&
+                Targets.SetEquals(other.Targets);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                return ((UsedWeapons?.GetHashCode() ?? 0) * 397) ^
+                    (Targets?.GetHashCode() ?? 0);
+            }
+        }
     }
 }